lambda-packages/api-transformation/requirements.txt
confluent-kafka==1.9.2 ; python_full_version == "3.9.7"
python-dotenv==0.21.0 ; python_full_version == "3.9.7"
regex==2022.10.31 ; python_full_version == "3.9.7"
